+++ Optimizer/Program.cs
@@ -14,7 +14,9 @@ namespace Optimizer
         // Enter current version here
 
         internal readonly static float Major = 7;
-        internal readonly static float Minor = 1;
+        internal readonly static float Minor = 2;
+
+        internal readonly static bool EXPERIMENTAL_BUILD = false;
 
         internal static string GetCurrentVersionTostring()
         {